+++ regen/camera/cube-camera.h
@@ -2,7 +2,7 @@
 #define REGEN_CUBE_CAMERA_H_
 
 #include <regen/camera/camera.h>
-#include "regen/gl/blend-state.h"
+#include "regen/gl/states/blend-state.h"
 
 namespace regen {
 	/**
