+++ regen/gl/CMakeLists.txt
@@ -1,9 +1,9 @@
 
 # list of files in this directory
-file(GLOB SOURCES *.cpp queries/*.cpp)
-file(GLOB HEADER_FILES *.h queries/*.h)
+file(GLOB SOURCES *.cpp queries/*.cpp states/*.cpp)
+file(GLOB HEADER_FILES *.h queries/*.h states/*.h)
 
 install(
     FILES ${HEADER_FILES}
-    DESTINATION ${HEADER_INSTALL_PATH}/gl-types
+    DESTINATION ${HEADER_INSTALL_PATH}/gl
 )
