+++ regen/gl/states/alpha-state.cpp
