+++ regen/gl/states/stencil-state.h
