+++ regen/objects/terrain/grass-patch.cpp
@@ -1,7 +1,7 @@
 #include "grass-patch.h"
 #include "regen/textures/texture-loader.h"
 #include "regen/objects/composite-mesh.h"
-#include "regen/gl/alpha-state.h"
+#include "../../gl/states/alpha-state.h"
 
 using namespace regen;
