+++ regen/textures/fbo-state.cpp
@@ -1,5 +1,5 @@
 #include "regen/utility/string-util.h"
-#include "regen/gl/atomic-states.h"
+#include "regen/gl/states/atomic-states.h"
 
 #include "fbo-state.h"
 #include "screen-state.h"
